Update db.lua
Thanks to wow-tools and wowhead, I was able to review all spells in the game now.
I removed 105 irrelevant NPCs and added 102 new NPCs. 

More changes:
- Disorient and Charm abilities are now added.

We're down to 993 relevant spells (from the 1863 spells). This is because I now take only the first occurrence of all linked abilities.
Even though wowhead has been the best source for this exercise, it's imperfect. There's some oddities we'll only be able to correct if you report it.
For example, some charge spells result in a knockback and stun, and some do not. Yet, all charge spells are now classified as stun and knockback.
In the theme of being cautious, this dataset is good enough and errs on the side of caution.